Created two new projects.
ApeSupport, the primary project for decompilation and nicification of APE.

OLD_ApeDecompile is an IronPython project containing the reference decompiler for APE. It's badly documented, and it's been added so I can write up documentation as I work through it. Once that's done, I can start porting it step by step.
I'm guessing about two to three days of work to get the documentation and possibly port completely sorted out.

Once APE is done, the next step is MD2 - if APE hurt my face, MD2 will melt it. After that, we're looking at some solid work on BSP and Planet. After that... not so much a problem.