Added the complete Irrlicht source.
Should've committed much sooner.

I've made a mess of changes to the underlying code in Irrlicht in an attempt to compile it as C++/CLI application.

The greatest challenge (at the moment) is compiling all the dependencies as C++/CLI. There are a lot of non-compliant code (method definitions, externs, typedefs) that I'm trying to sort out in a haphazard, yet careful, manner.